+++ common/sql_model.go
@@ -1,0 +1,9 @@
+package common
+
+import "time"
+
+type SQLModel struct {
+	Id        int        `json:"id"`
+	CreatedAt *time.Time `json:"created_at"`
+	UpdatedAt *time.Time `json:"updated_at"`
+}
