+++ module/item/storage/sql.go
@@ -1,0 +1,11 @@
+package storage
+
+import "gorm.io/gorm"
+
+type sqlStore struct {
+	db *gorm.DB
+}
+
+func NewSQLStore(db *gorm.DB) *sqlStore {
+	return &sqlStore{db: db}
+}
